% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipu.R
\name{cipu}
\alias{cipu}
\title{Confidence interval for the population proportion when the population proportion is UNKNOWN}
\usage{
cipu(x, c)
}
\arguments{
\item{x}{A vector of values in a distribution}

\item{c}{A value for the confidence level in whole number form...no decimals}
}
\value{
A confidence interval for the population proportion
}
\description{
Confidence interval for the population proportion when the population proportion is UNKNOWN
}
\examples{
cipu(dataset, 95)
cipu(dataset, 99)
}

