% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cixk.R
\name{cixk}
\alias{cixk}
\title{Confidence interval for the population mean when standard deviation is KNOWN}
\usage{
cixk(x, sigma, c)
}
\arguments{
\item{x}{A vector of values in a distribution}

\item{sigma}{A value for the population standard deviation}

\item{c}{A value for the confidence level in whole number form...no decimals}
}
\value{
A confidence interval for the population mean
}
\description{
Confidence interval for the population mean when standard deviation is KNOWN
}
\examples{
cixk(Orange$age, 920, 95)
cixk(Orange$age, 920, 99)
}

