% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cixu.R
\name{cixu}
\alias{cixu}
\title{Confidence interval for the population mean when standard deviation is UNKNOWN}
\usage{
cixu(x, c)
}
\arguments{
\item{x}{A vector of values in a distribution}

\item{c}{A value for the confidence level in whole number form...no decimals}
}
\value{
A confidence interval for the population mean
}
\description{
Confidence interval for the population mean when standard deviation is UNKNOWN
}
\examples{
cixu(Orange$age, 95)
cixu(Orange$age, 99)
}

